requirements.txt
streamlit
pymongo 
pillow 
google-generativeai
